Keep or compute SHA1 hash based on what is passed
In the `toSvg()` function, if the hash value is a hex string of at
least length 11, then keep using that value as the hash. Else compute
a SHA1 hash on the value passed. This is consistent with how the
javascript lib does it.

Update the API docs to match the javascript lib docs and this behavior.

The unit tests for `toSvg()` have six SVG files that represent input
hash strings that are already valid hex strings greater than length 11,
or some that are not and will be hashed with SHA1. Those six test SVG
files were generated manually and visually compared to the SVG images
produced by the javascript lib via the jdenticon.com website. There you
can type in the input value and see the resultimg image. For example:

    https://jdenticon.com/#icon-deadbeef123

The image produced at the above URL was manually (visually) confirmed
to match the SVG file we generated with jdenticon-kotlin, saved at:

    test/kotlin/resources/svg-output/deadbeef123.svg